adds script that builds board graph
using a factory function and recursion I have theoretically created a grapph linking every square to it's possible moves.

At this point it seems to work but I haven't thought of a way to test that every permutation is available.

Next steps are to create a script that can traverse the graph and find the fastest route between squares.