fixes social graph creation
The previous version of this code attempted to create a circular graph that would link everysquare with all of their possible moves and vice versa.
This became difficult to manage as I could not find a way to carry over the previous nodes across recursions and then apply them to the current. Or prevent the code from creating duplicates of every node.

Instead I have hard coded all of the squares and put them into a binary tree. Then using the tree I created duplicate squares and linked them to their tree counterparts. Lastly I created code that will traverse the tree and fill in the missing square connections.